backend/requirements.txt
fastapi
uvicorn
elasticsearch
python-dotenv
